+++ packages/store/src/base-store.ts
@@ -61,8 +61,12 @@ export abstract class BaseStore<
   private readonly historyDriver: StoreHistoryDriver<TData>;
 
   /** @inheritDoc */
-  readonly travelTo = (index: number): void =>
-    this.historyDriver.travelTo(index);
+  readonly restoreStoreAt = (index: number): void =>
+    this.historyDriver.restoreStoreAt(index);
+
+  /** @inheritDoc */
+  readonly restoreResource = <K extends StoreKey<TData>>(key: K, index?: number): void =>
+    this.historyDriver.restoreResource(key, index);
 
   /** @inheritDoc */
   readonly undo = (): boolean => this.historyDriver.undo();
@@ -168,6 +172,8 @@ export abstract class BaseStore<
     this.historyDriver = createStoreHistory<TData>({
       captureSnapshot: () => consumer.createSnapshot(),
       applySnapshot: (snapshot) => consumer.applySnapshot(snapshot),
+      applyKeyUpdate: (key, snapshotState) => consumer.applyKeyUpdate(key, snapshotState),
+      getAllKeys: () => this.storeKeys,
       applyMessage: (message) => consumer.applyMessage(message),
       channel: options?.channel,
     });