+++ samples/taskflurry/src/app/projects/presentation/list/projects-list.component.ts
@@ -4,7 +4,7 @@ import { RouterLink } from '@angular/router';
 import { ProjectsFacade } from '../../application/facades/projects.facade';
 import { Project } from '../../domain/models/project.model';
 import { PROJECTS_STORE } from '../../application/store/projects.store';
-import { StoreHistoryVizComponent } from '../../../core/devtools/store-history-viz.component';
+import { HistoryVizStore, StoreHistoryVizComponent } from '../../../core/devtools/store-history-viz.component';
 
 @Component({
   selector: 'app-projects-list',
@@ -45,7 +45,7 @@ export class ProjectsListComponent implements OnInit {
   });
   protected readonly isLoadingTasks = computed(() => this.tasksState().isLoading ?? false);
 
-  protected readonly historyStore = inject(PROJECTS_STORE);
+  protected readonly historyStore = inject(PROJECTS_STORE) as unknown as HistoryVizStore;
 
   ngOnInit(): void {
     this.facade.loadProjects();